refactored out footer and nav from app
